Fix playback in browsers using H264Stream
As suggested by mneil on videojs's Slack channel. Thanks!

Runs the input NALs through H264Stream to manipulate it a bit,
which seems to get things in the proper order and details. If
this is a bottleneck in future it can be optimized away, but
this is fine for now!

Fixes https://github.com/brion/raw-to-mse/issues/1